Save 16 bits in common case
This draft has a bigger tag than draft-ietf-avt-srtp-ekt-03.  Here is a way
to reduce the size by 16 bits.  Note that the SPI looses one bit, but it is
already described in this draft as too small in the relevant use-case so
one bit less does not make this worse.  If this really is an issue then
yet another bit could similarly be defined to indicate a 61 bits SPI.